nvim: rename mini_conf to mini
