+++ config/nvim/init.lua
@@ -98,7 +98,7 @@ end, {})
 -- Set up 'mini.deps' (customize to your liking)
 require('mini.deps').setup({ path = { package = path_package } })
 
-require('mini_conf')
+require('mini')
 require('lsp')
 require('plugins')
 require('oplog')
