+++ config/nvim/lua/mini/editor.lua
