+++ config/nvim/lua/mini/nav.lua
